+++ app/models/grouping.rb
@@ -728,6 +728,23 @@ class Grouping < ApplicationRecord
     next_grouping
   end
 
+  def get_random_incomplete(current_role)
+    if current_role.ta?
+      # Get relevant groupings for a TA
+      results = self.assignment.current_results.joins(grouping: :tas).where(
+        marking_state: Result::MARKING_STATES[:incomplete],
+        'roles.id': current_role.id
+      )
+
+    else
+      # Get all groupings in an assignment -- typically for an instructor
+      results = self.assignment.current_results.where(
+        marking_state: Result::MARKING_STATES[:incomplete]
+      )
+    end
+    results.where.not('groupings.id': self.id).order('RANDOM()').first&.grouping
+  end
+
   private
 
   def add_assignment_folder(group_repo)